+++ Test/Microsoft.ApplicationInsights.Test/Shared/TelemetryClientTest.cs
@@ -458,7 +458,9 @@
 
             var timestamp = DateTimeOffset.Now;
 #pragma warning disable CS0612 // Type or member is obsolete
+#pragma warning disable CS0618 // Type or member is obsolete
             client.TrackDependency("name", "command name", timestamp, TimeSpan.FromSeconds(42), false);
+#pragma warning restore CS0618 // Type or member is obsolete
 #pragma warning restore CS0612 // Type or member is obsolete
 
             var dependency = (DependencyTelemetry)sentTelemetry.Single();
@@ -988,6 +990,63 @@
             client.Track(eventTelemetry);
 
             Assert.AreEqual("test", eventTelemetry.Context.Internal.SdkVersion);
+        }
+
+        [TestMethod]
+        public void TrackClearsTelemetryContextRawStorageTempAfterInitializersAreRun()
+        {
+            const string keyTemp = "fooTemp";
+            const string detailTemp = "barTemp";
+            const string keyPerm = "fooPerm";
+            const string detailPerm = "barPerm";
+            var sentTelemetry = new List<ITelemetry>();
+            var channel = new StubTelemetryChannel
+            {
+                OnSend = t =>
+                {
+                    // Upon reaching TelemetryChannel temp object should not exist.
+                    Assert.IsFalse(t.Context.TryGetRawObject(keyTemp, out object detailTmp));
+
+                    // Upon reaching TelemetryChannel perm object should remain.
+                    Assert.IsTrue(t.Context.TryGetRawObject(keyPerm, out object detailPrm));
+                }
+            };           
+            var configuration = new TelemetryConfiguration(string.Empty, channel);
+
+            configuration.TelemetryProcessorChainBuilder.Use((next) => new StubTelemetryProcessor(next)
+            {
+                OnProcess = t =>
+                {
+                
+                // Upon reaching TelemetryProcessor temp object should not exist.
+                Assert.IsFalse(t.Context.TryGetRawObject(keyTemp, out object detailTmp));
+
+                // Upon reaching TelemetryProcessor perm object should remain.
+                Assert.IsTrue(t.Context.TryGetRawObject(keyPerm, out object detailPrm));
+                }
+            });
+
+            configuration.TelemetryProcessorChainBuilder.Build();
+            
+            var telemetryInitializer = new StubTelemetryInitializer();
+            telemetryInitializer.OnInitialize = item =>
+            {
+                // TelemetryInitializer should be able to access both temp and perm objects.
+                Assert.IsTrue(item.Context.TryGetRawObject(keyTemp, out object detailTmp));
+                Assert.IsTrue(item.Context.TryGetRawObject(keyPerm, out object detailPrm));
+            };
+
+            configuration.TelemetryInitializers.Add(telemetryInitializer);
+
+            var client = new TelemetryClient(configuration);
+            var telemetry = new StubTelemetry();
+            telemetry.Context.StoreRawObject(keyTemp, detailTemp, true);
+            telemetry.Context.StoreRawObject(keyPerm, detailPerm, false);
+            
+            // Calling Track will in turn call all TelemetryInitializers followed by TelemetryProcessors
+            // and Channel. Test here is to validate that temp RawObjects stored in TelemetryContext gets
+            // cleared after Initialiers are run.
+            client.Track(telemetry);
         }
 
         #endregion