+++ Test/ServerTelemetryChannel.Test/TelemetryChannel.Nuget.Tests/TelemetryChannelTests.cs
@@ -1,5 +1,7 @@
 ﻿namespace Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel
 {
+    using System;
+    using System.Diagnostics;
     using System.Linq;
     using System.Xml.Linq;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -12,10 +14,10 @@
         {
             string emptyConfig = ConfigurationHelpers.GetEmptyConfig();
             XDocument configAfterTransform = ConfigurationHelpers.InstallTransform(emptyConfig);
-
+            Trace.WriteLine(configAfterTransform.ToString());
             var typeToFind = typeof(ServerTelemetryChannel);
 
-            var node = ConfigurationHelpers.GetTelemetryChannel(configAfterTransform)
+            var node = ConfigurationHelpers.GetTelemetryChannelFromDefaultSink(configAfterTransform)
                 .FirstOrDefault(element => element.Attribute("Type").Value == ConfigurationHelpers.GetPartialTypeName(typeToFind));
 
             Assert.IsNotNull(node);
@@ -25,11 +27,10 @@
         public void UninstallRemovesTelemetryChannel()
         {
             string emptyConfig = ConfigurationHelpers.GetEmptyConfig();
-            XDocument configAfterInstall = ConfigurationHelpers.InstallTransform(emptyConfig);
-
+            XDocument configAfterInstall = ConfigurationHelpers.InstallTransform(emptyConfig);            
             XDocument configAfterUninstall = ConfigurationHelpers.UninstallTransform(configAfterInstall.ToString());
-
-            Assert.AreEqual(0, ConfigurationHelpers.GetTelemetryChannel(configAfterUninstall).ToList().Count);
+            Trace.WriteLine(configAfterUninstall.ToString());
+            Assert.AreEqual(0, ConfigurationHelpers.GetTelemetryChannelFromDefaultSink(configAfterUninstall).ToList().Count);
         }
     }
 }