+++ src/Microsoft.ApplicationInsights/Channel/Transmission.cs
@@ -26,7 +26,6 @@
         private static readonly TimeSpan DefaultTimeout = TimeSpan.FromSeconds(100);
 #if NETSTANDARD1_3
         private static HttpClient client = new HttpClient() { Timeout = System.Threading.Timeout.InfiniteTimeSpan };
-        private CancellationTokenSource cancellationTokenSource;
 #endif
         private int isSending;
 
@@ -37,17 +36,17 @@
         {
             if (address == null)
             {
-                throw new ArgumentNullException("address");
+                throw new ArgumentNullException(nameof(address));
             }
 
             if (content == null)
             {
-                throw new ArgumentNullException("content");
+                throw new ArgumentNullException(nameof(content));
             }
 
             if (contentType == null)
             {
-                throw new ArgumentNullException("contentType");
+                throw new ArgumentNullException(nameof(contentType));
             }
 
             this.EndpointAddress = address;
@@ -92,6 +91,7 @@
             private set;
         }
 
+#pragma warning disable CA1819 // "Properties should not return arrays" - part of the public API and too late to change.
         /// <summary>
         /// Gets the content of the transmission.
         /// </summary>
@@ -100,6 +100,7 @@
             get;
             private set;
         }
+#pragma warning restore CA1819 // "Properties should not return arrays" - part of the public API and too late to change.
 
         /// <summary>
         /// Gets the content's type of the transmission.
@@ -161,8 +162,11 @@
                 using (MemoryStream contentStream = new MemoryStream(this.Content))
                 {
                     HttpRequestMessage request = this.CreateRequestMessage(this.EndpointAddress, contentStream);
-                    this.cancellationTokenSource = new CancellationTokenSource(this.Timeout);             
-                    await client.SendAsync(request, this.cancellationTokenSource.Token).ConfigureAwait(false);
+
+                    using (var ct = new CancellationTokenSource(this.Timeout))
+                    {
+                        await client.SendAsync(request, ct.Token).ConfigureAwait(false);
+                    }
                     
                     return null;
                 }