+++ src/Microsoft.ApplicationInsights/DataContracts/TraceTelemetry.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.Globalization;
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility;
     using Microsoft.ApplicationInsights.Extensibility.Implementation;
@@ -133,6 +134,12 @@
         public ITelemetry DeepClone()
         {
             return new TraceTelemetry(this);
+        }
+
+        /// <inheritdoc/>
+        public void SerializeData(ISerializationWriter serializationWriter)
+        {
+            serializationWriter.WriteProperty(this.Data);            
         }
 
         /// <summary>