+++ src/ServerTelemetryChannel/Implementation/TransmissionSender.cs
@@ -43,7 +43,7 @@
             {
                 if (value < 0)
                 {
-                    throw new ArgumentOutOfRangeException("value");
+                    throw new ArgumentOutOfRangeException(nameof(value));
                 }
 
                 this.capacity = value;